// web/apps/webapp/src/queries/__generated__/customerFloorPlanView_availableResources_query.graphql.ts
export type customerFloorPlanView_availableResources_query$data = {
  readonly availableResources: ReadonlyArray<{
    readonly customTags: ReadonlyArray<{
      readonly color: string | null | undefined;
      readonly name: string | null | undefined;
      readonly uniqueId: string;
    }>;
    readonly name: string;
    readonly uniqueId: string;
    readonly zones: ReadonlyArray<{
      readonly color: string | null | undefined;
      readonly name: string | null | undefined;
      readonly uniqueId: string;
    }>;
  }>;
  readonly " $fragmentType": "customerFloorPlanView_availableResources_query";
};
export type customerFloorPlanView_availableResources_query$key = {
  readonly " $data"?: customerFloorPlanView_availableResources_query$data;
  readonly " $fragmentSpreads": FragmentRefs<"customerFloorPlanView_availableResources_query">;
};